+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- Data types: `str`, `int`, `float`
 - Simple arithmetic
 - Variable assignment
 - Print to console